fix(network): 修复RESP解析器空bulk字符串处理和数据类型问题
- 在Redis Null Bulk String特殊处理后重置解析器状态并返回成功
- 在完成命令解析后重置解析器状态以避免状态污染
- 将数组大小和bulk长度的数据类型从size_t改为long long以支持负值处理

+++ include/network/RESPParser.h
@@ -35,8 +35,8 @@ class RESPParser {
 
  private:
   State state_;
-  size_t array_size_;  // 数组总元素数
-  size_t bulk_len_;    // 当前正在解析的字符串长度
+  long long array_size_;  // 数组总元素数
+  long long bulk_len_;    // 当前正在解析的字符串长度
 };
 
 #endif  // NOVAKV_RESPPARSER_H
